input: touchscreen: added support for LDC 3001 based touchscreen
The LP101WX2 10" display panel has an integrated touchscreen
controller. This controller, under the covers has a LDC3001
ROIC and a LGD STM32F103 MCU. It is multi-touch capable  and can
tracking up to ten touches simultenously. The firmare for the
MCU fetches data from the ROIC over SPI. After proccessing, the
controller passes on the tocuh information via an I2C interface.

This is a simple driver processing the i2c messages and providing
linux multi-touch compatible input events to the user space.

Change-Id: I16693c67e5fd3380bb86ffbefec2008f3e0e6ae0
Signed-off-by: Subramaniam Chanderashekarapuram <subramaniam.ca@ti.com>
Signed-off-by: Marcus Cooksey <mcooksey@ti.com>