gpio: mcp23s08: Set initial direction and value from DT
GPIO framework does not allow to set a set of gpios is one shot.
Even the gpio_request_array iterates over each entry and sets up gpio
lines one by one.

For some devices (e.g. FPDlink deserializer), if the gpios are used as
to drive some of the control pins, each combination of the control
pins define a specific behavior, some other combination may lead to
inconsistent state, so it is needed to set all of the gpio pins in one go.

This can be achieved by reading the initial directions and values
from device tree. Add support for reading the device tree properties
and initialize the gpio pins as specified by the DT.

Change-Id: I54d01396dc1b345c9423e3f1976e3cc14d1e6b6e
Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>