remoteproc/omap: add "late attach" support
Add the necessary support in the OMAP remoteproc driver to enable the
"late attach" functionality. The 'late attach' functionality has the
processors already loaded and running by the time the kernel has booted.

The 'late attach' support in OMAP remoteproc driver is added through
minimal changes like skipping the releasing of the processor reset.
The processor reset is skipped as this relies on the omap_hwmod /
omap_device API which perform a module disable/enable sequence in
addition to the reset programming. Other required functionality is
achieved through the late attach support in the remoteproc driver core.

During late attach, we use non-zeroing dma ops to prevent the kernel
from overwriting already loaded code and data segments. When shutting
down the processor, we restore the normal zeroing dma ops.  This allows
the kernel to clear memory when loading a new remoteproc binary or
during error recovery with the current remoteproc binary.

The 'late attach' capability is announced through a "ti,late-attach"
property on the respective remoteproc node in the device tree blob.
The corresponding IOMMU and timer nodes should also have this property
set. All the nodes should also have the "ti,no-idle-on-init" and
"ti,no-reset-on-init" properties set to prevent omap_hwmod from
resetting and idling these devices on boot.

The parsed property is stored in a .late_attach field in the rproc
data structure. The field is used for differentiating a normal boot
from a late-attach boot. The "ti,late-attach" feature is relevant only
for the first time boot of the kernel, and is therefore removed from
the remoteproc device-tree node on the first probe. This prevents the
.late_attach flag from being set for subsequent probes of the remoteproc
node. The flag is also reset during the stopping of the remoteproc,
so that any error recovery results in a regular boot.

Change-Id: If52e3e3980fe0bc76725ef94c0ffce415cfee138
Signed-off-by: Robert Tivy <rtivy@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Venkateswara Rao Mandela <venkat.mandela@ti.com>
Signed-off-by: Angela Stegmaier <angelabaker@ti.com>