arm: dtsi: dra7: Create example dts file for using DMA pool for late attach
For late attach case, it is desireable to make the memory
as "no-map" so that it is not used by the kernel for
any other purpose. When the reserved mem is defined as
"reusable", then it is possible that the kernel may
allocate that memory before the remoteproc is probed
and allocates the memory. In this case, the memory can
be corrupted. Making the memory as "no-map" removes
this possibility.

Also use the "ti,late-attach-dma-pool" flag for the mmu
node so that the iommu module knows that the dma-pool
is being used and can take certain actions based on this
related to the page table address.

Please note that when using DMA pools, the allocation scheme
is different than with CMA pools, and the size of the carveout
may need to be increased or alignment changed depending on the
memory needs of the image being loaded. Also, the vmalloc size
may need increasing, depending on system configuration, number
of remote cores using a DMA pool, and the sizes of the DMA pools.
This is because the DMA pool is not accessible in kernel linear
memory and needs to be mapped to be accessed.

Change-Id: I3a2f16e15980b7939f7fd7543b3ad924410d42fe
Signed-off-by: Angela Stegmaier <angelabaker@ti.com>