drivers: dma-coherent: Add Non-Zeroing Alloc Function
Add a new function, dma_malloc_from_coherent(), which
takes a parameter that specifies if the allocated memory
should be zero-ed or not. Modify dma_alloc_from_coherent
to call dma_malloc_from_coherent with the flag set to
true in order to zero the memory. In this way
dma_alloc_from_coherent behaves the same as before.

Change-Id: I7c497fc88739116e233736e9d9a95ecbd849ec8f
Signed-off-by: Angela Stegmaier <angelabaker@ti.com>