ARM: dma-mapping: Call the non-zeroing dma-coherent alloc from malloc function
The malloc call, which is intented to be the same as alloc,
except that the memory is not zero-ed, should also not
zero memory that is allocated using dma_alloc_from_coherent().
This patch changes the call to dma_alloc_from_coherent() to
dma_malloc_from_coherent() with the zero flag set to false
in order to prevent zero-ing of the memory.

Change-Id: I4b7fea40d7dcd4443a5a3b1c9476b67da40998ff
Signed-off-by: Angela Stegmaier <angelabaker@ti.com>