CPSW: Add fixed-link DT binding support
This commit adds a "fixed-link" binding for the CPSW slave property
that can be used instead of a dedicated PHY with "phy_id". This
provides better support for CPSW slaves that are connected to a
dedicated MAC that does not have an associated PHY and always remains
fixed at a constant configuration.

Change-Id: Ic8cfd5361898ac0982ffdb0e55b2f95f012534f9
Signed-off-by: J.D. Schroeder <jay.schroeder@garmin.com>
Signed-off-by: Sriramakrishnan A G <srk@ti.com>