iommu/omap: Update iommu to work with ioremap pagetable
In the late attach case, the page table is taken from the end
of the reserved memory instead of the usual method of allocating.
In the case where the reserved mem is a DMA pool rather than a
CMA pool, the usual method of translating physical addresses to
virtual addresses and vice-versa does not work. The virtual address
retrieved in that case would be inaccessible. So, in this case, it
is required to ioremap the page table so that it is accessible
in the kernel. As a result of this, the phys_to_virt that is used
during iommu_enable will not work, because phys_to_virt doesn't
work for ioremap memory. So, at the point where the ioremap is
performed, we are now saving the physical address, which is later
used during iommu_enable rather than doing a virt_to_phys.

A new property "ti,late-attach-dma-pool" is added to let the iommu
know that a dma pool is being used so that the proper method of
address translation can be used.

This patch makes the changes to ioremap the pagetable in the case
of late attach with a dma pool, save the physical address of the
pagetable during the attach phase, and then use the saved physical
address during iommu_enable. With these changes, the iommu works
properly with late attach using a DMA pool.

Change-Id: I40ad5c74ba2a27dd9c5141128c2386ae2c5fe52d
Signed-off-by: Angela Stegmaier <angelabaker@ti.com>