mmc: host: omap_hsmmc: Fix disabled mmc clock in SDR50
The mmc clock was accidentally disabled by the tuning function for
the SDR50 mode when max clock is > 52MHz.  The mmc clock is expected
to be temporarily disabled as part of the tuning sequence, but it is
later enabled again.

The driver performs two checks to determine if tuning required:
- Clock is > 52MHz
- CAPA2 register indicates that tuning is needed for SDR50

The driver was disabling the mmc clock between the two checks, but
didn't restore it if the second check failed (i.e. when tuning is not
required for SDR50).

The tuning function is reworked so that the cases when tuning is
required are tested before disabling the mmc clock.

Change-Id: I1c29fbac7eaefb797781c7d4930f92e8e783dc6a
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>