usb: xhci: i896: add quirk flag for broken PED bits
this patch is for fixing

i896: USB xHCI port disable feature does not work

picked from ti-linux-4.1.y

commit 5dea9519038a ("usb: xhci: add quirk flag for broken PED bits")

Some devices from Texas Instruments suffer from
a silicon bug where Port Enabled/Disabled bit
should not be used to silence an erroneous device.

The bug is so that if port is disabled with PED
bit, an IRQ for device removal (or attachment)
will never fire.

Just for the sake of completeness, the actual
problem lies with SNPS USB IP and this affects
all known versions up to 3.00a. A separate
patch will be added to dwc3 to enabled this
quirk flag if version is <= 3.00a.

Signed-off-by: Felipe Balbi <balbi@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>
[append public errata ID and pull to 3.14 auto baseline]
Change-Id: Ie431876759baf753f22618e4341793c155e4b5f8
Signed-off-by: Praneeth Bajjuri <praneeth@ti.com>