dra7: dt: update regulator constraints for main VDDs
MPU,DSP,IVA,GPU and CORE voltage domains support
AVS class 0, so the voltage that needs to be configured
actually is read from efuses which may vary between a min and
a max, to allow this variation an absolute-max-voltage
has been introduced to allow the regulator to set
any voltage between the efused value and the max.
Update regulator constraints to meet this requirement.

Change-Id: I76b2d7726ec91e96576a1bb325f184277bebb273
Signed-off-by: Ravikumar Kattekola <rk@ti.com>