arm: dts: dra72: Remove pinmux from kernel
With iodelay configuration and pad configuration being performed
in isolation from bootloader, all the pinmux config is to be moved
to the bootloader. Hence removing the pinmux entries and corresponding
pinctrl properties in all peripheral device nodes.

Note that the updated bootloader with patches for pad configuration
is needed for this kernel to be functional.

Change-Id: Ib0b2a43aac94f1ef57ec3c1a6d2cde4ecc733cbb
Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>