drm/omap: flush crtcs only when needed
At the moment omapdrm flushes all crtcs when any client is closed. But
as only the master client can queue new work for crtcs, we can do the
flush only for the master.

We also do the flush in master_drop, so that a master_drop followed by
device close will still cause a flush.

Change-Id: Ie0bebed735a1b654122c5d84f7c52b69ba54229c
Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>