drm/omap: call waiter->notify from non-atomic context
The SGX driver is called via the sync_op notify callback, and this
callback may happen in atomic context. However, the SGX driver uses
mutex in the callback, causing the kernel to crash.

The best fix would probably be to change the SGX driver. However, this
patch fixes the issue also by first collecting the notify callbacks
which need to be called, and then actually calling them a bit later from
non-atomic context.

Change-Id: I58c42dbf84f4d97cb73f78667d7a2a88213b3946
Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>