drm/omap: fix kfree call when freeing omap_obj
omapdrm allocates struct omap_gem_objects, but when freeing them, it
uses the pointer to struct drm_gem_object. These are currently the same,
so no issue is seen, but if the struct omap_gem_object is changed,
problems could arise.

So fix the possible future problem by kfree'ing the objects using the
correct pointer.

Change-Id: Ia645a7ad7ed348ecad39be7afe293d8b4667363c
Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>