media: ti-vpe: cal: Restrict DMA to avoid memory corruption
When setting DMA for video capture from CSI channel, if the DMA size
is not given, it ends up writing as much data as sent by the camera.

This may lead to overwriting the buffers causing memory corruption.
Observed green lines on the default framebuffer.

Restrict the DMA to maximum height as specified in the S_FMT ioctl.

Change-Id: I3bba24c213960e93da1f66dd2ade8f5bebd2c887
Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>