media: ti-vpe: cal: Perform dummy read after the PHY reset
According to the TRM, CSI2 D-Phy module gets the clock only when
CSI2 PHY registers are accessed. But the PHY needs 3 SCP cycles to
come out of reset. Therefore, a dummy software read is needed to
complete the PHY reset sequence.

Add a dummy read after the PHY reset and before the PHY registers are
accessed. This avoid reading junk data when configuring the PHY module.

Change-Id: I562bcaf44da668356307ae6bf794e65ab2285b8f
Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>