net/rpmsg: handle callback for deleted sockets
In the case where a callback comes in for a socket that has
already been deleted, a kernel panic will result when the
invalid (NULL) sk is accessed. This could happen, for example,
when an application is incorrectly terminated while still
communicating with the remote core. The remote core is unaware
of the application's termination and continues to send messages.

This patch adds a check in the callback for a NULL (invalid) sk
before accesses the socket, and throws a warning if a callback
for an invalid socket is recevied. This will prevent the kernel
panic from happening, but note that this will note notify the remote
core of the error and it may still be unaware of the removal of the
HOST rpmsg-proto socket unless it has some other mechanism for
checking that the socket is still alive.

Change-Id: Ib3d128d6e1343a390ef7b0180ce94508b7739b1d
Signed-off-by: Angela Stegmaier <angelabaker@ti.com>