dwc3: gadget: shortpkt termination not handled during data stage
when host sends data during control write request and
if there is shortpacket the data phase must be terminated
don't issue the ep0 request to receive next data.

This fixes the usb3.0 device does not get detect when
connected to windows host. When windows host send
SET_SEL control write request to set the exit latency,
with shortpacket data, during data phase the shortpkt was
not handled and causing the enumeration failure.

Change-Id: I33d738ff2c9207d2f348aa5b2fa8b5bd4d5ce887
Signed-off-by: Ravi Babu <ravibabu@ti.com>