media: ti-vpe: cal: One time phy initialization
The CAL Phy reset state machine depends on the external peripheral
to complete the reset.

Only when the LPHS transition is detected, the reset is completed.
For this, multiple options have been tried:
- Toggle the TX clock at stop and start
- Toggle the CTRL module bits to connect disconnect PHY signals
- Disable and Enable the CTRL module after the reset sequence is issued

But none of this works reliably and subsequent capture invocation may
or may not work.
Therefore, its better to configure the Phy only one time and only
setup/disable the DMAs upon capture start/stop.

Change-Id: I755b4ea9a0dc0c2dc5049fdaa62d83d286ca2863
Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>