ASoC: davinci-mcasp: Removing shared DAI ISR
If a McASP DAI is shared with another core (e.g. a DSP), all
interrupts are to be handled by the remote core and should not be
serviced by the davinci-mcasp driver.

This will prevent a flood of interrupts to the MPU in case of a
remoteproc reload of the core owning the shared DAI as the remote
core will not have a chance to properly reset the McASP.

Change-Id: Iaa3fb4a75ac70e20a7c102e9a8a771c87d3de7b2
Signed-off-by: Stephen Molfetta <sjmolfetta@ti.com>