arm: dts: dra72: vision: Remove pinmux from kernel
With iodelay configuration and pad configuration being performed
in isolation from bootloader, all the pinmux config is to be moved
to the bootloader. Hence removing the pinmux entries and corresponding
pinctrl properties in all peripheral device nodes.

Note that the updated bootloader with patches for pad configuration
is needed for this kernel to be functional.

Change-Id: Ia685d244c524338ed0713e61a07da98ab19ef2e2
Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>