arm: dts: dra7xx: enable ldc3001 touchscreen
The LDC 3001 based touchscreen present in the
LP101WX2 panel is connected as an i2c1 slave and
has its interrupt line hooked up to the wakueup2 pin.
The wakeup2 pin is therefore programmed to gpio pin 2
in bank 1 and acts as the interrupt source for the
touchscreen interrupt. Since the i2c slave address of
the audio codec hooked up to the i2c1 bus clashes with
the touchscreen controller address, the address of
the audio codec has been modified to be 0x19 instead
of 0x18. Also the atmel touchscreen device tree data is removed.

Change-Id: I86cdb45613af19209b43f85c28c40b180dacafed
Signed-off-by: Marcus Cooksey <mcooksey@ti.com>