ARM: dts: dra7*-evm: Use consistent rates for ATL clocks
There are two ATL instances that are being used for radio:
 - ATL1 for DSP rendering through McASP6
 - ATL2 for ARM rendering through McASP3

The BCLK rate required for McASP6 is 11.2896MHz, while McASP3
is using a lower BCLK rate.  The McASP3 BCLK can be run at the
same rate as McASP6's in order to make the BCLK rate transparent
for the radio software and independent of the type of rendering
mechanism.

Change-Id: Id7a2b83d4976a3fbc198bb9221ebe5d40bde78b6
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>