ti_config_fragments: android: Use long ASoC DAPM power-down time
Due to a board limitation, the line-out and headphone outputs
in the DRA7 EVM's CPU board are prone to audio pops/clicks
artifacts when powering up the audio codec.

ASoC DAPM changes the bias level of the codec dynamically
based on the codec usage, which would ultimately cause the
audio artifacts described above. The workaround consists of
keeping the codec always on by using the longest power-down
time allowed by DAPM in order to minimize those artifacts.

Change-Id: Ibbbc803614a958232f31401a8a32ba1901bc535c
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>