ARM: dts: dra7xx-clocks: Set GPU DPLL and its output clock rates
The GPU DPLL in DRA7xx provides the output clocks for only the GPU
subsystem in DRA7xx.

This DPLL is currently not configured by bootloader. Use the DT
standard properties "assigned-clocks" and "assigned-clock-rates" to
set the GPU DPLL clock rate and the rates for its derivative clocks
at boot time to avoid any functional issues caused when using the
GPU subsystem if the DPLL is not locked. The DPLL will automatically
transition into a low-power stop mode when the associated output
clocks are not utilized or gated automatically.

The clock rates are chosen based on the OPP_NOM values defined as per
recommendation from design team. The DPLL locked frequency is kept at 1277 MHz,
so that the value for the divider clock, dpll_gpu_m2_ck, can be set to 425.67 MHz
for OPP_NOM.

This patch also sets the assigned-clock-parents properties for
gpu_hyd_gclk_mux and gpu_hyd_gclk_mux so that these clocks are derived
from dpll_gpu_m2_ck instead of dpll_core_h14x2_ck

Based on Logic similar to
Commit a30be71ab192 ("ARM: dts: dra7xx-clocks: Set IVA DPLL and its output clock rates")

Signed-off-by: Subhajit Paul <subhajit_paul@ti.com>
[upload to gerrit and minor commit msg tweak]
Change-Id: Iddf5b1220c13f917e3ed235b829ea39acb5a26d6
Signed-off-by: Praneeth Bajjuri <praneeth@ti.com>