rpmsg: return error from create_channel if endpoint creation fails
In the device_register call for rpmsg_create_channel, the device
probe is called, which attempts to create an endpoint. If that
endpoint creation fails, the probe call fails, but device_register
does not propagate that error. The rpmsg_create_channel call
needs to explicitly check that the endpoint creation was successful
after the device_register call, and return an error in case
of failure.

Change-Id: I20c27eddb4ddfd55c2fde250917d156f41c145e9
Signed-off-by: Angela Stegmaier <angelabaker@ti.com>