rpmsg: Call the non-zeroing allocation function
For DMA pool scenario the plugged in allocation ops are not
used for carveout allocation. It is necessary to call a differnt
function that does not zero the DMA pool memory gotten from
the call to dma_alloc_coherent.

Call the dma_malloc_coherent (non-zeroing) api in the case
of late attach to guarantee that the memory is not zero-ed for
either a CMA or DMA pool.

Change-Id: I87f5e3d7fb892212d7737e52eca3a315a5ef6590
Signed-off-by: Angela Stegmaier <angelabaker@ti.com>