remoteproc/omap: consolidate instance-specific pre_shutdown ops
The .pre_shutdown() ops is defined to handle certain pre-shutdown
acvities. This is used by the DSP remoteproc devices on DRA7 SoCs
to recover properly during error recovery when an active EDMA
transaction was happening. These ops are currently defined to be
instance specific, consolidate them into a single instance and
derive the instance-id based on the device string names. This
reduces the duplicity in pdata definitions and handling the
same in remoteproc code.

Change-Id: I648cb13596290dfeed7d79a3d4126a7af493d95a
Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Sheng Zhao <shengzhao@ti.com>